video-chat-app-with-django/videochat/requirements.txt
agora-token-builder==1.0.0
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
eventlet==0.37.0
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
imbalanced-learn==0.12.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mysqlclient==2.2.4
numpy==1.23.5
opencv-python==4.6.0.66
packaging==24.1
panda==0.3.1
pandas==2.2.3
Pillow==9.3.0
platformdirs==4.3.6
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.4.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
wsproto==1.2.0
